refactor: rename segment module to row_group
